Notes:
* blenderbuttons and ICON_SNAP_PEEL_OBJECT were not merged.

+++ source/gameengine/Expressions/EmptyValue.cpp
@@ -76,7 +76,7 @@ this object
 
 
 
-float CEmptyValue::GetNumber()
+double CEmptyValue::GetNumber()
 {
 	return 0;
 }
